first

+++ Task1.py
@@ -1,4 +1,4 @@
-num1=int(input("Enter the base of the triangle: "))
-num2=int(input("Enter the height of the triangle: "))
+num1=float(input("Enter the base of the triangle: "))
+num2=float(input("Enter the height of the triangle: "))
 area=(0.5*num1*num2)
 print(area)
